Moved soapProperties to commons
